reorganize code + delete legacy directory
decided i don't want to maintain that -- anyone using < v3 can just stay on that version
