+++ src/common/circular-dependency-error.ts
