+++ src/common/guard.ts
