+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from "./api";
 export * from "./common";
 export * from "./lifecycle";
-export * from "./utils";
