--- src/legacy/container/index.ts
@@ -1,1 +1,0 @@
-export * from "./container"
