--- src/legacy/decorators/index.ts
@@ -1,2 +1,0 @@
-export * from './inject.decorator'
-export * from './injectable.decorator'
