+++ src/lifecycle/index.ts
@@ -1,5 +1,5 @@
 export * from "./destructible";
-export * from "./is-on-destroy.guard";
-export * from "./is-on-init.guard";
-export * from "./on-destroy.interface";
-export * from "./on-init.interface";
+export * from "./is-on-destroy";
+export * from "./is-on-init";
+export * from "./on-destroy";
+export * from "./on-init";
