+++ src/lifecycle/is-on-destroy.ts
@@ -1,5 +1,5 @@
-import { Guard } from "../utils";
-import { OnDestroy } from "./on-destroy.interface";
+import { Guard } from "../common";
+import { OnDestroy } from "./on-destroy";
 
 export const isOnDestroy: Guard<OnDestroy> = (x: unknown): x is OnDestroy => {
 	return typeof x["onDestroy"] === "function";
