--- src/utils/constants.ts
@@ -1,10 +1,0 @@
-/**
- * @author kuro <kuro@kuroi.io>
- * @namespace kuroi.io.Syringe.Utils.Constants
- */
-
-export namespace Constants {
-
-	export const GLOBAL_SCOPE: string = 'global'
-
-}
